+++ src/widgets/header/Header.stories.tsx
@@ -83,7 +83,7 @@ export const AuthWithLanguageSelector: Story = {
 export const LogoLink: Story = {
   args: { variant: 'guest' },
   play: async ({ canvas }) => {
-    const logo = canvas.getByRole('link', { name: 'remarkgram' })
+    const logo = canvas.getByRole('link', { name: 'Inctagram' })
 
     await expect(logo).toHaveAttribute('href', '/')
   },